converts alert to ts
